star/requirements.txt
rasterio~=1.3.2
numpy~=1.22.4
geopandas~=0.11.1
pandas~=1.4.3
matplotlib~=3.5.2
altair~=4.2.0
plotly~=5.9.0
folium~=0.12.1.post1
sklearn~=0.0
scikit-learn~=1.1.1
streamlit~=1.13.0
openpyxl~=3.0.10
pillow~=9.2.0
streamlit-multipage~=0.0.18
ogr~=0.40.0
streamlit-aggrid~=0.3.3
tensorflow-estimator~=2.9.0
tensorflow
keras
keras-preprocessing~=1.1.2
